requirements.txt
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
httplib2==0.21.0
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.21.9
requests==2.28.1
requests-oauthlib==1.3.1
